fix comment

+++ kv_options.h
@@ -9,10 +9,10 @@
 
 namespace eloqstore
 {
-constexpr int KB = 1024;
-constexpr int MB = KB * 1024;
-constexpr int GB = MB * 1024;
-constexpr int64_t TB = static_cast<int64_t>(GB) * 1024;
+constexpr int64_t KB = 1024;
+constexpr int64_t MB = KB * 1024;
+constexpr int64_t GB = MB * 1024;
+constexpr int64_t TB = GB * 1024;
 
 constexpr uint8_t max_overflow_pointers = 128;
 constexpr uint16_t max_read_pages_batch = max_overflow_pointers;
